User-settings, new cursors, and rewrote script
Rewrote script to make the disable hotkey instant and also better
readability
Updated script to work for Flash Projector, Official Website,
Kongregate, and Steam
Added user-defined settings for the cursor-type and manual-disable
hotkey.
Modified cursors to have an odd number of pixels to allow the cursor
"hotspot" to be in the dead center.
Created new circle-based cursors and a thinner rainbow cursor.